requirements.txt
Django
django-angular
djangorestframework==3.8.2
coreapi
django-cors-headers